analysis_flow/requirements.txt
# =============================================================================
# KRA-ORA Medical Analysis System v2.0 - Backend Dependencies
# Python 3.10+ required
# =============================================================================

# Core ML and NLP
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.2
accelerate>=0.24.0

# Vector Database & Supabase
supabase>=2.3.0
faiss-cpu>=1.7.4

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0

# Text Processing
langchain>=0.1.0
langchain-community>=0.0.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
structlog>=23.2.0

# Email Validation
email-validator
